com.utilities.buildpipeline 1.5.2 (#40)
- fix iOS validation when using swift frameworks

## Copilot Summary

This pull request includes changes to improve the iOS build process and
update the package version. The most important changes include
modifications to the `OnPostProcessBuild` method in `IOSBuildInfo.cs` to
handle Swift standard libraries and bitcode settings, and an update to
the package version in `package.json`.

### Improvements to iOS Build Process:
*
[`Utilities.BuildPipeline/Packages/com.utilities.buildpipeline/Editor/Platforms/IOSBuildInfo.cs`](diffhunk://#diff-693f7ac2a58668190af6c0273485e1aafa8e08fd54ee96e46a5e2f9df9507d9eL18-R38):
Modified the `OnPostProcessBuild` method to set
`ALWAYS_EMBED_SWIFT_STANDARD_LIBRARIES` for the main target and Unity
framework target, and updated the handling of `ENABLE_BITCODE` for
different Unity versions.

### Package Version Update:
*
[`Utilities.BuildPipeline/Packages/com.utilities.buildpipeline/package.json`](diffhunk://#diff-1bf8683ef5b05ae61aeafadc407894b0756b6409f8d12585eedf92f815487e3eL6-R6):
Updated the package version from `1.5.1` to `1.5.2`.